requirements.txt
PyQt6
sounddevice
requests
PyObjC
pynput
pyinstaller
python-dotenv
